+++ BetterTriggers/Models/EditorData/ExplorerElementFolder.cs
@@ -98,7 +98,7 @@ namespace BetterTriggers.Models.EditorData
             }
         }
 
-        public int GetId()
+        public long GetId()
         {
             throw new NotImplementedException();
         }
