+++ BetterTriggers/Models/EditorData/ExplorerElementRoot.cs
@@ -84,7 +84,7 @@ namespace BetterTriggers.Models.EditorData
             }
         }
 
-        public int GetId()
+        public long GetId()
         {
             throw new NotImplementedException();
         }
