+++ BetterTriggers/Models/SaveableData/Trigger.cs
@@ -1,11 +1,13 @@
 ﻿using BetterTriggers.Containers;
+using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 
 namespace BetterTriggers.Models.SaveableData
 {
     public class Trigger : IReferable
     {
-        public int Id;
+        public long Id;
         public string Comment;
         public bool IsScript;
         public bool RunOnMapInit;
